better tests for invites
