+++ packages/pds/tests/account.test.ts
@@ -1,37 +1,17 @@
 import { once, EventEmitter } from 'events'
-import AtpAgent, {
-  ComAtprotoServerCreateAccount,
-  ComAtprotoServerResetPassword,
-} from '@atproto/api'
+import AtpAgent, { ComAtprotoServerResetPassword } from '@atproto/api'
 import { IdResolver } from '@atproto/identity'
 import * as crypto from '@atproto/crypto'
 import Mail from 'nodemailer/lib/mailer'
 import { AppContext, Database } from '../src'
 import * as util from './_util'
 import { ServerMailer } from '../src/mailer'
-import { DAY } from '@atproto/common'
-import { genInvCodes } from '../src/api/com/atproto/server/util'
 
 const email = 'alice@test.com'
 const handle = 'alice.test'
 const password = 'test123'
 const passwordAlt = 'test456'
 const minsToMs = 60 * 1000
-
-const createInviteCode = async (
-  agent: AtpAgent,
-  uses: number,
-  forUser?: string,
-): Promise<string> => {
-  const res = await agent.api.com.atproto.server.createInviteCode(
-    { useCount: uses, forUser },
-    {
-      headers: { authorization: util.adminAuth() },
-      encoding: 'application/json',
-    },
-  )
-  return res.data.code
-}
 
 describe('account', () => {
   let serverUrl: string
@@ -47,9 +27,6 @@ describe('account', () => {
 
   beforeAll(async () => {
     const server = await util.runTestServer({
-      inviteRequired: true,
-      userInviteInterval: DAY,
-      userInviteEpoch: Date.now() - 3 * DAY,
       termsOfServiceUrl: 'https://example.com/tos',
       privacyPolicyUrl: '/privacy-policy',
       dbPostgresSchema: 'account',
@@ -79,20 +56,9 @@ describe('account', () => {
     }
   })
 
-  let inviteCode: string
-
-  it('creates an invite code', async () => {
-    inviteCode = await createInviteCode(agent, 1)
-    const split = inviteCode.split('-')
-    const host = split.slice(0, -2).join('.')
-    const code = split.slice(-2).join('-')
-    expect(host).toBe('pds.public.url') // Hostname of public url
-    expect(code.length).toBe(11)
-  })
-
   it('serves the accounts system config', async () => {
     const res = await agent.api.com.atproto.server.describeServer({})
-    expect(res.data.inviteCodeRequired).toBe(true)
+    expect(res.data.inviteCodeRequired).toBe(false)
     expect(res.data.availableUserDomains[0]).toBe('.test')
     expect(typeof res.data.inviteCodeRequired).toBe('boolean')
     expect(res.data.links?.privacyPolicy).toBe(
@@ -106,21 +72,8 @@ describe('account', () => {
       email: 'bad-handle@test.com',
       handle: 'did:bad-handle.test',
       password: 'asdf',
-      inviteCode,
     })
     await expect(promise).rejects.toThrow('Input/handle must be a valid handle')
-  })
-
-  it('fails on bad invite code', async () => {
-    const promise = agent.api.com.atproto.server.createAccount({
-      email,
-      handle,
-      password,
-      inviteCode: 'fake-invite',
-    })
-    await expect(promise).rejects.toThrow(
-      ComAtprotoServerCreateAccount.InvalidInviteCodeError,
-    )
   })
 
   let did: string
@@ -131,7 +84,6 @@ describe('account', () => {
       email,
       handle,
       password,
-      inviteCode,
     })
     did = res.data.did
     jwt = res.data.accessJwt
@@ -151,13 +103,11 @@ describe('account', () => {
   })
 
   it('allows a custom set recovery key', async () => {
-    const inviteCode = await createInviteCode(agent, 1)
     const recoveryKey = (await crypto.EcdsaKeypair.create()).did()
     const res = await agent.api.com.atproto.server.createAccount({
       email: 'custom-recovery@test.com',
       handle: 'custom-recovery.test',
       password: 'custom-recovery',
-      inviteCode,
       recoveryKey,
     })
 
@@ -171,7 +121,6 @@ describe('account', () => {
   })
 
   it('allows a user to bring their own DID', async () => {
-    const inviteCode = await createInviteCode(agent, 1)
     const userKey = await crypto.Secp256k1Keypair.create()
     const handle = 'byo-did.test'
     const did = await ctx.plcClient.createDid({
@@ -191,7 +140,6 @@ describe('account', () => {
       handle,
       did,
       password: 'byo-did-pass',
-      inviteCode,
     })
 
     expect(res.data.handle).toEqual(handle)
@@ -199,7 +147,6 @@ describe('account', () => {
   })
 
   it('requires that the did a user brought be correctly set up for the server', async () => {
-    const inviteCode = await createInviteCode(agent, 1)
     const userKey = await crypto.Secp256k1Keypair.create()
     const baseDidInfo = {
       signingKey: ctx.repoSigningKey.did(),
@@ -216,7 +163,6 @@ describe('account', () => {
       email: 'byo-did@test.com',
       handle: 'byo-did.test',
       password: 'byo-did-pass',
-      inviteCode,
     }
 
     const did1 = await ctx.plcClient.createDid({
@@ -313,7 +259,6 @@ describe('account', () => {
   })
 
   it('disallows duplicate email addresses and handles', async () => {
-    const inviteCode = await createInviteCode(agent, 2)
     const email = 'bob@test.com'
     const handle = 'bob.test'
     const password = 'test123'
@@ -321,7 +266,6 @@ describe('account', () => {
       email,
       handle,
       password,
-      inviteCode,
     })
 
     await expect(
@@ -329,7 +273,6 @@ describe('account', () => {
         email: email.toUpperCase(),
         handle: 'carol.test',
         password,
-        inviteCode,
       }),
     ).rejects.toThrow('Email already taken: BOB@TEST.COM')
 
@@ -338,19 +281,16 @@ describe('account', () => {
         email: 'carol@test.com',
         handle: handle.toUpperCase(),
         password,
-        inviteCode,
       }),
     ).rejects.toThrow('Handle already taken: bob.test')
   })
 
   it('disallows improperly formatted handles', async () => {
-    const inviteCode = await createInviteCode(agent, 1)
     const tryHandle = async (handle: string) => {
       await agent.api.com.atproto.server.createAccount({
         email: 'john@test.com',
         handle,
         password: 'test123',
-        inviteCode,
       })
     }
     await expect(tryHandle('did:john')).rejects.toThrow(
@@ -391,66 +331,19 @@ describe('account', () => {
     await expect(tryHandle('atp.test')).rejects.toThrow('Reserved handle')
   })
 
-  it('fails on used up invite code', async () => {
-    const promise = agent.api.com.atproto.server.createAccount({
-      email: 'bob@test.com',
-      handle: 'bob.test',
-      password: 'asdf',
-      inviteCode,
-    })
-    await expect(promise).rejects.toThrow(
-      ComAtprotoServerCreateAccount.InvalidInviteCodeError,
-    )
-  })
-
-  it('handles racing invite code uses', async () => {
-    const inviteCode = await createInviteCode(agent, 1)
-    const COUNT = 10
-
-    let successes = 0
-    let failures = 0
-    const promises: Promise<unknown>[] = []
-    for (let i = 0; i < COUNT; i++) {
-      const attempt = async () => {
-        try {
-          await agent.api.com.atproto.server.createAccount({
-            email: `user${i}@test.com`,
-            handle: `user${i}.test`,
-            password: `password`,
-            inviteCode,
-          })
-          successes++
-        } catch (err) {
-          failures++
-        }
-      }
-      promises.push(attempt())
-    }
-    await Promise.all(promises)
-    expect(successes).toBe(1)
-    expect(failures).toBe(9)
-  })
-
   it('handles racing signups for same handle', async () => {
     const COUNT = 10
 
-    const invite1 = await createInviteCode(agent, COUNT)
-    const invite2 = await createInviteCode(agent, COUNT)
-
     let successes = 0
     let failures = 0
     const promises: Promise<unknown>[] = []
     for (let i = 0; i < COUNT; i++) {
       const attempt = async () => {
         try {
-          // Use two invites to ensure per-invite locking doesn't
-          // give the appearance of fixing a race for handle.
-          const invite = i % 2 ? invite1 : invite2
           await agent.api.com.atproto.server.createAccount({
             email: `matching@test.com`,
             handle: `matching.test`,
             password: `password`,
-            inviteCode: invite,
           })
           successes++
         } catch (err) {
@@ -638,156 +531,5 @@ describe('account', () => {
         password,
       }),
     ).resolves.toBeDefined()
-  })
-
-  it('allow users to get available user invites', async () => {
-    // first pretend account was made 2 days in the past
-    const twoDaysAgo = new Date(Date.now() - 2 * DAY).toISOString()
-    await ctx.db.db
-      .updateTable('user_account')
-      .set({ createdAt: twoDaysAgo })
-      .where('did', '=', did)
-      .execute()
-    const res1 = await agent.api.com.atproto.server.getAccountInviteCodes()
-    expect(res1.data.codes.length).toBe(2)
-
-    // use both invites and confirm we can't get any more
-    await ctx.db.db
-      .insertInto('invite_code_use')
-      .values(
-        res1.data.codes.map((code) => ({
-          code: code.code,
-          usedBy: 'did:example:test',
-          usedAt: new Date().toISOString(),
-        })),
-      )
-      .execute()
-    const res2 = await agent.api.com.atproto.server.getAccountInviteCodes()
-    expect(res2.data.codes.length).toBe(2)
-  })
-
-  it('creates invites based on epoch', async () => {
-    // pretend account was made 10 days ago
-    const tenDaysAgo = new Date(Date.now() - 10 * DAY).toISOString()
-    await ctx.db.db
-      .updateTable('user_account')
-      .set({ createdAt: tenDaysAgo })
-      .where('did', '=', did)
-      .execute()
-
-    // we have a 3 day epoch so should still get 1 code
-    const res = await agent.api.com.atproto.server.getAccountInviteCodes({
-      includeUsed: false,
-    })
-    expect(res.data.codes.length).toBe(1)
-    const res2 = await agent.api.com.atproto.server.getAccountInviteCodes()
-    expect(res2.data.codes.length).toBe(3)
-
-    // we pad their account with some additional used codes from the past which should not change anything
-    const inviteRows = genInvCodes(ctx.cfg, 10).map((code) => ({
-      code: code,
-      availableUses: 1,
-      disabled: 0 as const,
-      forUser: did,
-      createdBy: did,
-      createdAt: new Date(Date.now() - 5 * DAY).toISOString(),
-    }))
-    await ctx.db.db.insertInto('invite_code').values(inviteRows).execute()
-    await ctx.db.db
-      .insertInto('invite_code_use')
-      .values(
-        inviteRows.map((row) => ({
-          code: row.code,
-          usedBy: 'did:example:test',
-          usedAt: new Date().toISOString(),
-        })),
-      )
-      .execute()
-
-    const res3 = await agent.api.com.atproto.server.getAccountInviteCodes({
-      includeUsed: false,
-    })
-    expect(res3.data.codes.length).toBe(1)
-  })
-
-  it('prevents use of disabled codes', async () => {
-    const first = await createInviteCode(agent, 1)
-    const accntCodes =
-      await agent.api.com.atproto.server.getAccountInviteCodes()
-    const second = accntCodes.data.codes[0].code
-
-    // disabled first by code & second by did
-    await agent.api.com.atproto.admin.disableInviteCodes(
-      {
-        codes: [first],
-        accounts: [did],
-      },
-      {
-        headers: { authorization: util.adminAuth() },
-        encoding: 'application/json',
-      },
-    )
-
-    const attempt = async (code: string) => {
-      await agent.api.com.atproto.server.createAccount({
-        email: 'disable@test.com',
-        handle: 'disable.test',
-        inviteCode: code,
-        password: 'disabled',
-      })
-    }
-
-    await expect(attempt(first)).rejects.toThrow(
-      ComAtprotoServerCreateAccount.InvalidInviteCodeError,
-    )
-    await expect(attempt(second)).rejects.toThrow(
-      ComAtprotoServerCreateAccount.InvalidInviteCodeError,
-    )
-  })
-
-  it('does not allow disabling all admin codes', async () => {
-    const attempt = agent.api.com.atproto.admin.disableInviteCodes(
-      {
-        accounts: ['admin'],
-      },
-      {
-        headers: { authorization: util.adminAuth() },
-        encoding: 'application/json',
-      },
-    )
-    await expect(attempt).rejects.toThrow('cannot disable admin invite codes')
-  })
-
-  it('creates many invite codes', async () => {
-    const accounts = ['did:example:one', 'did:example:two', 'did:example:three']
-    const res = await agent.api.com.atproto.server.createInviteCodes(
-      {
-        useCount: 2,
-        codeCount: 2,
-        forAccounts: accounts,
-      },
-      {
-        headers: { authorization: util.adminAuth() },
-        encoding: 'application/json',
-      },
-    )
-    expect(res.data.codes.length).toBe(3)
-    const fromDb = await ctx.db.db
-      .selectFrom('invite_code')
-      .selectAll()
-      .where('forUser', 'in', accounts)
-      .execute()
-    expect(fromDb.length).toBe(6)
-    const dbCodesByUser = {}
-    for (const row of fromDb) {
-      expect(row.disabled).toBe(0)
-      expect(row.availableUses).toBe(2)
-      dbCodesByUser[row.forUser] ??= []
-      dbCodesByUser[row.forUser].push(row.code)
-    }
-    for (const { account, codes } of res.data.codes) {
-      expect(codes.length).toBe(2)
-      expect(codes.sort()).toEqual(dbCodesByUser[account].sort())
-    }
   })
 })
